requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-SQLAlchemy==2.5.1
Werkzeug==2.2.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
